+++ libs/parsing/index/src/lib/auto-complete/completion-for/add-completion-special-functions.interface.ts
@@ -1,3 +1,5 @@
+import { TOKEN_NAMES } from '@idl/tokenizer';
+
 /**
  * Regex to detect auto complete for special functions
  */
@@ -12,3 +14,11 @@ export const ENVI_TASK_REGEX = /^envi.+task$/i;
  * Regex to just detect IDL Task function calls
  */
 export const IDL_TASK_REGEX = /^idl.+task$/i;
+
+/**
+ * Track tokens that we can do task completion for
+ */
+export const ALLOWED_SPECIAL_COMPLETION: { [key: string]: undefined } = {};
+ALLOWED_SPECIAL_COMPLETION[TOKEN_NAMES.QUOTE_SINGLE] = undefined;
+ALLOWED_SPECIAL_COMPLETION[TOKEN_NAMES.QUOTE_DOUBLE] = undefined;
+ALLOWED_SPECIAL_COMPLETION[TOKEN_NAMES.STRING_TEMPLATE_LITERAL] = undefined;