+++ libs/parsing/index/src/lib/auto-complete/completion-for/add-completion-special-procedures.interface.ts
@@ -1,0 +1,4 @@
+/**
+ * Regex to detect auto complete for special procedures
+ */
+export const SPECIAL_PROCEDURE_REGEX = /call_procedure/i;