+++ CHANGELOG.md
@@ -16,6 +16,15 @@ Document some advanced types so users may try them out and provide feedback. The
 
 - Read more in the extension documentation
 
+## Unreleased
+
+Added layer controls to the Notebook Map.
+
+- Generic title for each layer (raster/vector)
+- Slider to make transparent or not
+
+Notebook maps now listen to theme changes from VSCode and, when swapping between light and dark themes, the basemap will update to match the theme of VSCode (light map for light theme; dark map for dark theme)
+
 ## 4.6.0 - August 2024
 
 Fixed an issue where we open a file that we have stopped in, even if it is compiled as a SAVE file. Now, only PRO files will be jumped to.